pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "polaroid-png"
version = "0.1.0"
description = "Backward-compatible PNG extension with embedded back face + metadata."
readme = "README.md"
requires-python = ">=3.9"
dependencies = ["pillow>=10.0.0"]

authors = [
  {name = "Xinq", email = "dabufan@gmail.com"}
]
license = {text = "MIT"}
keywords = ["png", "polaroid", "image-format", "lossless"]
classifiers = [
  "License :: OSI Approved :: MIT License",
  "Programming Language :: Python :: 3",
  "Topic :: Multimedia :: Graphics"
]

[project.scripts]
polaroid-encode = "polaroid_png.cli:encode_main"
polaroid-decode = "polaroid_png.cli:decode_main"
